Split out docker-compose specific code into "dev backend".
Long run this opens up the possibility of using a local K3S cluster, minikube
or even a kull K8S install as a dev backend.
