+++ kubetools_client/dev/__main__.py
@@ -4,7 +4,12 @@ from kubetools_client.dev import dev
 from kubetools_client.main import run_cli
 
 # Import click command groups
-from kubetools_client.dev import container, environment, logs, scripts  # noqa: F401, I100
+from kubetools_client.dev import (  # noqa: F401, I100, I202
+    container,
+    environment,
+    logs,
+    scripts,
+)
 
 
 run_cli(dev)
