+++ kubetools_client/dev/container.py
@@ -3,7 +3,7 @@ import os
 import click
 
 from . import dev
-from .container_util import (
+from .backend import (
     build_containers,
     exec_container,
     get_container_status,
