+++ kubetools_client/settings.py
@@ -18,6 +18,7 @@ class KubetoolsSettings(object):
     DEV_DEFAULT_ENV = 'dev'  # default environment name in dev
     DEV_HOST = 'localhost'  # dev host to link people to (should map to 127.0.0.1)
     DEV_CONFIG_DIRNAME = '.kubetools'  # project directroy to generate compose config
+    DEV_BACKEND = 'docker_compose'  # backend to use for development
 
     KUBETOOLS_SESSION = None  # user auth/session details (generated by the server)
